+++ CHANGELOG.md
@@ -5,7 +5,7 @@ We are currently working on porting this changelog to the specifications in
 This project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 1.3.0 - Unreleased
+## Version 1.3.0 - 2022-06-14
 
 ### Changed
 * Removed experimental warning from `Path.copy` and `Path.move`; they are now well enough tested.